backend/requirements.txt
fastapi==0.115.7
llama_index==0.12.14
pydantic==2.10.6
python-dotenv==1.0.1
numpy==1.24.4
scipy==1.11.3
